add framework, WIP
